HAJAE-7 add custom hook: typing effect
