+++ src/app/globals.scss
@@ -1,3 +1,4 @@
+@import '@/styles/animations.scss';
 @import '@/styles/colors.scss';
 @import '@/styles/inputs.scss';
 @import '@/styles/typography.scss';
